I've written the basline of the project.
Now we can randomize a network based on a matrix of probabilities containing single and double links. Next steps are: addapt to bipartite networks (which should be pretty straight forward), write C++ version of pglmm and convert distance matrix to probabilities. Another important point would be breaking down functions to be callable for the user...