This is the beta version of the package.
It works. However, tests are still required and I should go deeper into the fitting problems (find cases in which we can't find a good fitting for B and check whether or not we can identify them before). In addition, we will need to restructure the package and change the following:
- check functions should just check
- implement other randomizations
- simplify the way the input need to be introduced
- break down the main function into different pieces so we don't need to run all the tests every time
- restructure the output of the program
- write the stupid documentation